Add files via upload
Copy of the functions and the DicePairArray used for the final project. You can add this to a Google colab instance and then use "from cantstopfuncs import * " in your code to be able to use all these functions. (see e.g. https://stackoverflow.com/questions/48905127/importing-py-files-in-google-colab ... use the updated instructions from Nov 2018!)